Vis kp2d update

See merge request zeotrope/zoehuman/mmhuman3d!23

+++ mmhuman3d/core/visualization/visualize_keypoints2d.py
@@ -16,7 +16,7 @@ from .keypoint_utils import search_limbs
 def plot_kp2d_frame(kp2d_person: np.ndarray,
                     canvas: np.ndarray,
                     limbs: dict = SMPLX_LIMBS_INDEX,
-                    palette: Union[dict, np.ndarray] = SMPLX_PALETTE,
+                    palette: Optional[Union[dict, np.ndarray]] = None,
                     draw_bbox: bool = False,
                     with_number: bool = False,
                     font_size: Union[float, int] = 0.5) -> np.ndarray:
@@ -38,7 +38,7 @@ def plot_kp2d_frame(kp2d_person: np.ndarray,
                 Defaults to 0.5.
 
     Returns:
-        np.ndarray: cv2 image of shape (H * W * 3).
+        np.ndarray: opencv image of shape (H * W * 3).
     """
     if kp2d_person.shape[-1] == 3:
         kp2d_person = kp2d_person[..., :-1]
@@ -50,10 +50,12 @@ def plot_kp2d_frame(kp2d_person: np.ndarray,
         bbox = _get_bbox(kp2d_person, canvas, expand=True)
     else:
         bbox = None
+    if palette is None:
+        palette = SMPLX_PALETTE
     for part_name, part_limbs in limbs.items():
         scatter_points_index = list(
             set(np.array([part_limbs]).reshape(-1).tolist()))
-        if part_name == 'body':
+        if isinstance(palette, dict) and part_name == 'body':
             thickness = 2
             radius = 3
             color = np.random.randint(
@@ -71,7 +73,6 @@ def plot_kp2d_frame(kp2d_person: np.ndarray,
 
         for limb_index, limb in enumerate(part_limbs):
             limb_index = min(limb_index, len(color) - 1)
-
             cv2.line(
                 canvas,
                 tuple(kp2d_person[limb[0]].astype(np.int32)),
@@ -121,38 +122,44 @@ def _get_bbox(keypoint_np: np.ndarray,
     return np.asarray([x_min, x_max, y_min, y_max])
 
 
-def render_kp2d_to_video(
+def visualize_kp2d(
     kp2d: np.ndarray,
     output_path: str,
     frame_list: Optional[List[str]] = None,
     limbs: Optional[Union[np.ndarray, List[int]]] = None,
+    palette: Optional[Iterable[int]] = None,
     data_source: str = 'mmpose',
     mask: Optional[Union[list, np.ndarray]] = None,
     start: int = 0,
     end: int = -1,
+    force: bool = False,
     resolution: Optional[Union[Tuple[int, int], list]] = None,
     fps: Union[float, int] = 30,
     draw_bbox: bool = False,
     with_number: bool = False,
     pop_parts: Iterable[str] = [],
 ) -> NoReturn:
-    """Render 2d keypoints to a video.
+    """Visualize 2d keypoints to a video or into a folder of frames.
 
     Args:
         kp2d (np.ndarray): should be array of shape (f * J * 2)
                                 or (f * n * J * 2)]
-        output_path (str): output video path.
+        output_path (str): output video path or image folder.
         frame_list (List[str]): list of frame paths, if None, would be
                 initialized as white background.
         limbs (Optional[Union[np.ndarray, List[int]]], optional):
                 if not specified, the limbs will be searched by search_limbs,
                 this option is for free skeletons like BVH file.
                 Defaults to None.
+        palette (Iterable, optional): specified palette, three int represents
+                (B, G, R). Should be tuple of list.
+                Defaults to None.
         data_source (str, optional): data source type. Defaults to 'mmpose'.
         mask (Optional[Union[list, np.ndarray]], optional):
                 mask to mask out the incorrect points. Defaults to None.
         start (int, optional): start frame index. Defaults to 0.
         end (int, optional): end frame index. Defaults to -1.
+        force (bool, optional): [description]. Defaults to False.
         resolution (Optional[Union[Tuple[int, int], list]], optional):
                 (width, height) of the output video
                 will be the same size as the original images if not specified.
@@ -171,10 +178,14 @@ def render_kp2d_to_video(
     Returns:
         NoReturn.
     """
-    if not Path(output_path).parent.is_dir():
-        raise FileNotFoundError(
-            f'The output folder does not exist: {Path(output_path).parent}')
-    if not Path(output_path).suffix.lower() in ['.mp4']:
+    output_pathinfo = Path(output_path)
+    if not output_pathinfo.parent.is_dir():
+        raise FileNotFoundError(f'The output root folder does not exist:\
+                    {output_pathinfo.parent}')
+    if output_pathinfo.suffix == '':
+        if not output_pathinfo.is_dir():
+            os.mkdir(output_path)
+    elif not output_pathinfo.suffix.lower() in ['.mp4']:
         raise FileNotFoundError(
             f'The output file should be .mp4: {output_path}')
     if kp2d.ndim == 3:
@@ -187,17 +198,23 @@ def render_kp2d_to_video(
                     and Path(frame_path).suffix.lower() in ['.png']):
                 raise FileNotFoundError(
                     f'The frame should be .png: {frame_path}')
-
-    temp_folder = os.path.join('.',
-                               os.path.basename(output_path) + '_temp_images')
+    if output_pathinfo.suffix.lower() in ['.mp4']:
+        temp_folder = os.path.join(
+            '.',
+            os.path.basename(output_path) + '_temp_images')
+    else:
+        temp_folder = output_path
     os.makedirs(temp_folder, exist_ok=True)
     if limbs is not None:
-        limbs_palette, limbs_target = np.random.randint(
-            0, high=255, size=(1, 3), dtype=np.uint8),
-        {'body': limbs.tolist() if isinstance(limbs, np.ndarray) else limbs}
+        limbs_target, limbs_palette = {
+            'body': limbs.tolist() if isinstance(limbs, np.ndarray) else limbs
+        }, np.random.randint(
+            0, high=255, size=(1, 3), dtype=np.uint8)
     else:
         limbs_target, limbs_palette = search_limbs(
             data_source=data_source, mask=mask)
+    if palette:
+        limbs_palette = np.array(palette, dtype=np.uint8)[None]
     num_frame = min(len(frame_list), num_frame)
     end = (min(num_frame, end) + num_frame) % num_frame
     kp2d = kp2d[start:end + 1]
@@ -229,10 +246,23 @@ def render_kp2d_to_video(
                 draw_bbox=draw_bbox,
                 with_number=with_number,
                 font_size=0.5)
-        cv2.imwrite(
-            os.path.join(temp_folder, '%06d.png' % frame_index), image_array)
-    images_to_video(
-        input_folder=temp_folder,
-        output_path=output_path,
-        remove_raw_file=True,
-        fps=fps)
+        if output_pathinfo.suffix.lower() in ['.mp4']:
+            cv2.imwrite(
+                os.path.join(temp_folder, '%06d.png' % frame_index),
+                image_array)
+        else:
+            if Path(temp_folder).absolute() == Path(
+                    frame_list[frame_index]).parent.absolute() and (not force):
+                raise FileExistsError(
+                    f'{temp_folder} exists (set --force to overwrite).')
+            else:
+                cv2.imwrite(
+                    os.path.join(temp_folder,
+                                 Path(frame_list[frame_index]).name),
+                    image_array)
+    if output_pathinfo.suffix.lower() in ['.mp4']:
+        images_to_video(
+            input_folder=temp_folder,
+            output_path=output_path,
+            remove_raw_file=True,
+            fps=fps)
